add transaction, EMF, iterfaces, named query, validation, config,

+++ src/main/java/ru/isaev/demoserver/controllers/BookController.java
@@ -1,10 +1,12 @@
 package ru.isaev.demoserver.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.*;
 import ru.isaev.demoserver.model.Book;
 import ru.isaev.demoserver.service.BookService;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @RestController
@@ -13,7 +15,7 @@ public class BookController {
     private final BookService bookServiceImpl;
 
     @Autowired
-    public BookController(BookService bookService) {
+    public BookController(@Qualifier(value = "bookServiceImpl")BookService bookService) {
         this.bookServiceImpl = bookService;
     }
 
@@ -23,31 +25,37 @@ public class BookController {
     }
 
     @PostMapping("/book-create")
-    public List<Book> createBook (@RequestBody Book book){
+    public List<Book> createBook (@Valid @RequestBody Book book){
         bookServiceImpl.saveBookInRepo(book);
         return bookServiceImpl.showAllBookInRepo();
     }
 
     @DeleteMapping("/book-delete/{id}")
     public List<Book> deleteBook(@PathVariable("id") Short id){
+        if(id < 1 || id == null){
+            return null;
+        }
         bookServiceImpl.deleteBookByIDInRepo(id);
         return bookServiceImpl.showAllBookInRepo();
     }
 
     @GetMapping("/book-get/{id}")
     public Book updateBook(@PathVariable("id") Short id){
+        if(id < 1 || id == null){
+            return null;
+        }
         return bookServiceImpl.findByIDInRepo(id);
     }
 
     @PatchMapping("/book-update")
-    public Book updateBook(@RequestBody Book book){
+    public Book updateBook(@Valid @RequestBody Book book){
         bookServiceImpl.updateBookInRepo(book);
         return book;
     }
 
     @GetMapping("/book-search")
-    public List<Book> searchBook(@RequestParam("bookName") String request, Book book){
-        if (request == "") {
+    public List<Book> searchBook(@RequestParam("bookName") String request){
+        if (request == "" || request == null) {
             return null;
         }
         return bookServiceImpl.findAllByBookNameOrGenreOrAuthorInRepo(request);
